requirements.txt
# Performance Marketing Agent Dependencies

# Core framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0

# AI / LLM
openai>=1.0.0

# HTTP requests (used by multiple services)
requests>=2.31.0

# HTML parsing (brand scraper)
beautifulsoup4>=4.12.0

# Date utilities
python-dateutil>=2.8.2

# Logging
python-json-logger>=2.0.0

# Prompt management
pyyaml>=6.0

# Database migrations
alembic>=1.13.0

# Task scheduling
APScheduler>=3.10.0

# Testing
pytest>=7.0.0
pytest-flask>=1.3.0

# --- Optional: uncomment as you set up each integration ---

# Email automation
# sendgrid>=6.10.0

# WhatsApp / SMS
# twilio>=8.0.0

# Google Ads
# google-ads>=24.0.0

# Facebook Marketing API
# facebook-business>=19.0.0
